feat:servo-motor

+++ src/drivers/ServoMotor.cpp
@@ -1,4 +1,4 @@
-#include "drivers/ServoMotor.hpp"
+#include "../include/drivers/ServoMotor.hpp"
 #include <mbed.h>
 
 
@@ -15,6 +15,6 @@ void ServoMotor::setAngleDeg(float degrees) {
     } else if (degrees > MAX_ANGLE) {
         degrees = MAX_ANGLE;
     }
-    int pulseWidth_us = MIN_PULSE_WIDTH + (degrees / MAX_ANGLE) * (MAX_PULSE_WIDTH - MIN_PULSE_WIDTH);
+    float pulseWidth_us = MIN_PULSE_WIDTH + degrees / MAX_ANGLE * (MAX_PULSE_WIDTH - MIN_PULSE_WIDTH);
     pwm.pulsewidth_us(pulseWidth_us); // パルス幅を設定
 }
